Cfg-gate whole rand::seq::index module
