+++ src/seq/index.rs
@@ -7,8 +7,6 @@
 // except according to those terms.
 
 //! Low-level API for sampling indices
-use super::gen_index;
-#[cfg(feature = "alloc")]
 use alloc::vec::{self, Vec};
 use core::slice;
 use core::{hash::Hash, ops::AddAssign};
@@ -16,10 +14,9 @@ use core::{hash::Hash, ops::AddAssign};
 #[cfg(feature = "std")]
 use super::WeightError;
 use crate::distributions::uniform::SampleUniform;
-#[cfg(feature = "alloc")]
 use crate::distributions::{Distribution, Uniform};
 use crate::Rng;
-#[cfg(all(feature = "alloc", not(feature = "std")))]
+#[cfg(not(feature = "std"))]
 use alloc::collections::BTreeSet;
 #[cfg(feature = "serde1")]
 use serde::{Deserialize, Serialize};
@@ -30,7 +27,6 @@ use std::collections::HashSet;
 ///
 /// Multiple internal representations are possible.
 #[derive(Clone, Debug)]
-#[cfg(feature = "alloc")]
 #[cfg_attr(feature = "serde1", derive(Serialize, Deserialize))]
 pub enum IndexVec {
     #[doc(hidden)]
@@ -39,7 +35,6 @@ pub enum IndexVec {
     USize(Vec<usize>),
 }
 
-#[cfg(feature = "alloc")]
 impl IndexVec {
     /// Returns the number of indices
     #[inline]
@@ -90,7 +85,6 @@ impl IndexVec {
     }
 }
 
-#[cfg(feature = "alloc")]
 impl IntoIterator for IndexVec {
     type IntoIter = IndexVecIntoIter;
     type Item = usize;
@@ -105,7 +99,6 @@ impl IntoIterator for IndexVec {
     }
 }
 
-#[cfg(feature = "alloc")]
 impl PartialEq for IndexVec {
     fn eq(&self, other: &IndexVec) -> bool {
         use self::IndexVec::*;
@@ -122,7 +115,6 @@ impl PartialEq for IndexVec {
     }
 }
 
-#[cfg(feature = "alloc")]
 impl From<Vec<u32>> for IndexVec {
     #[inline]
     fn from(v: Vec<u32>) -> Self {
@@ -130,7 +122,6 @@ impl From<Vec<u32>> for IndexVec {
     }
 }
 
-#[cfg(feature = "alloc")]
 impl From<Vec<usize>> for IndexVec {
     #[inline]
     fn from(v: Vec<usize>) -> Self {
@@ -171,7 +162,6 @@ impl<'a> Iterator for IndexVecIter<'a> {
 impl<'a> ExactSizeIterator for IndexVecIter<'a> {}
 
 /// Return type of `IndexVec::into_iter`.
-#[cfg(feature = "alloc")]
 #[derive(Clone, Debug)]
 pub enum IndexVecIntoIter {
     #[doc(hidden)]
@@ -180,7 +170,6 @@ pub enum IndexVecIntoIter {
     USize(vec::IntoIter<usize>),
 }
 
-#[cfg(feature = "alloc")]
 impl Iterator for IndexVecIntoIter {
     type Item = usize;
 
@@ -203,7 +192,6 @@ impl Iterator for IndexVecIntoIter {
     }
 }
 
-#[cfg(feature = "alloc")]
 impl ExactSizeIterator for IndexVecIntoIter {}
 
 /// Randomly sample exactly `amount` distinct indices from `0..length`, and
@@ -228,7 +216,6 @@ impl ExactSizeIterator for IndexVecIntoIter {}
 /// to adapt the internal `sample_floyd` implementation.
 ///
 /// Panics if `amount > length`.
-#[cfg(feature = "alloc")]
 #[track_caller]
 pub fn sample<R>(rng: &mut R, length: usize, amount: usize) -> IndexVec
 where
@@ -269,33 +256,6 @@ where
             sample_rejection(rng, length, amount)
         }
     }
-}
-
-/// Randomly sample exactly `N` distinct indices from `0..len`, and
-/// return them in random order (fully shuffled).
-///
-/// This is implemented via Floyd's algorithm. Time complexity is `O(N^2)`
-/// and memory complexity is `O(N)`.
-///
-/// Returns `None` if (and only if) `N > len`.
-pub fn sample_array<R, const N: usize>(rng: &mut R, len: usize) -> Option<[usize; N]>
-where
-    R: Rng + ?Sized,
-{
-    if N > len {
-        return None;
-    }
-
-    // Floyd's algorithm
-    let mut indices = [0; N];
-    for (i, j) in (len - N..len).enumerate() {
-        let t = gen_index(rng, j + 1);
-        if let Some(pos) = indices[0..i].iter().position(|&x| x == t) {
-            indices[pos] = j;
-        }
-        indices[i] = t;
-    }
-    Some(indices)
 }
 
 /// Randomly sample exactly `amount` distinct indices from `0..length`, and
@@ -432,7 +392,6 @@ where
 /// The output values are fully shuffled. (Overhead is under 50%.)
 ///
 /// This implementation uses `O(amount)` memory and `O(amount^2)` time.
-#[cfg(feature = "alloc")]
 fn sample_floyd<R>(rng: &mut R, length: u32, amount: u32) -> IndexVec
 where
     R: Rng + ?Sized,
@@ -464,7 +423,6 @@ where
 /// performance in all cases).
 ///
 /// Set-up is `O(length)` time and memory and shuffling is `O(amount)` time.
-#[cfg(feature = "alloc")]
 fn sample_inplace<R>(rng: &mut R, length: u32, amount: u32) -> IndexVec
 where
     R: Rng + ?Sized,
@@ -483,6 +441,7 @@ where
 
 trait UInt: Copy + PartialOrd + Ord + PartialEq + Eq + SampleUniform + Hash + AddAssign {
     fn zero() -> Self;
+    #[cfg_attr(feature = "alloc", allow(dead_code))]
     fn one() -> Self;
     fn as_usize(self) -> usize;
 }
@@ -530,7 +489,6 @@ impl UInt for usize {
 ///
 /// This function  is generic over X primarily so that results are value-stable
 /// over 32-bit and 64-bit platforms.
-#[cfg(feature = "alloc")]
 fn sample_rejection<X: UInt, R>(rng: &mut R, length: X, amount: X) -> IndexVec
 where
     R: Rng + ?Sized,
@@ -555,7 +513,6 @@ where
     IndexVec::from(indices)
 }
 
-#[cfg(feature = "alloc")]
 #[cfg(test)]
 mod test {
     use super::*;
